+++ airflow/providers/google/cloud/triggers/datafusion.py
@@ -80,7 +80,7 @@ class DataFusionStartPipelineTrigger(BaseTrigger):
             },
         )
 
-    async def run(self) -> AsyncIterator["TriggerEvent"]:  # type: ignore[override]
+    async def run(self) -> AsyncIterator[TriggerEvent]:  # type: ignore[override]
         """Gets current pipeline status and yields a TriggerEvent"""
         hook = self._get_async_hook()
         while True: