+++ dev/breeze/src/airflow_breeze/commands/testing_commands_config.py
@@ -35,6 +35,7 @@ TESTING_PARAMETERS: dict[str, list[dict[str, str | list[str]]]] = {
                 "--mysql-version",
                 "--mssql-version",
                 "--integration",
+                "--github-repository",
             ],
         },
         {
@@ -70,6 +71,7 @@ TESTING_PARAMETERS: dict[str, list[dict[str, str | list[str]]]] = {
                 "--postgres-version",
                 "--mysql-version",
                 "--mssql-version",
+                "--github-repository",
             ],
         },
         {