+++ airflow/providers/amazon/CHANGELOG.rst
@@ -38,8 +38,6 @@ Breaking changes
 
   Removed deprecated parameter ``max_tries`` from the Athena & EMR hook & operators in favor of ``max_polling_attempts``.
 
-  Disabled deprecated behavior of switching to an empty aws connection ID on error. You can set it to None explicitly.
-
   Removed deprecated method ``waiter`` from emr hook in favor of the more generic ``airflow.providers.amazon.aws.utils.waiter.waiter``
 
   Removed deprecated unused parameter ``cluster_identifier`` from Redshift Cluster's hook method ``get_cluster_snapshot_status``
@@ -73,6 +71,8 @@ Features
 * ``add a stop operator to emr serverless (#30720)``
 * ``SqlToS3Operator - Add feature to partition SQL table (#30460)``
 * ``New AWS sensor — DynamoDBValueSensor (#28338)``
+* ``Add a "force" option to emr serverless stop/delete operator (#30757)``
+* ``Add support for deferrable operators in AMPP (#30032)``
 
 Bug Fixes
 ~~~~~~~~~
@@ -82,6 +82,7 @@ Bug Fixes
 * ``Add ability to override waiter delay in EcsRunTaskOperator (#30586)``
 * ``Add support in AWS Batch Operator for multinode jobs (#29522)``
 * ``AWS logs. Exit fast when 3 consecutive responses are returned from AWS Cloudwatch logs (#30756)``
+* ``Fix async conn for none aws_session_token (#30868)``
 
 Misc
 ~~~~
@@ -89,10 +90,17 @@ Misc
 * ``Remove @poke_mode_only from EmrStepSensor (#30774)``
 * ``Organize Amazon providers docs index (#30541)``
 * ``Remove duplicate param docstring in EksPodOperator (#30634)``
+* ``Update AWS EMR Cluster Link to use the new dashboard (#30844)``
+* ``Restore aiobotocore as optional dependency of amazon provider (#30874)``
 
 .. Below changes are excluded from the changelog. Move them to
    appropriate section above if needed. Do not delete the lines(!):
    * ``Decouple "job runner" from BaseJob ORM model (#30255)``
+   * ``Upgrade ruff to 0.0.262 (#30809)``
+   * ``fixes to system tests following obsolete cleanup (#30804)``
+   * ``restore fallback to empty connection behavior (#30806)``
+   * ``Prepare docs for adhoc release of providers (#30787)``
+   * ``Prepare docs for ad-hoc release of Amazon provider (#30848)``
 
 7.4.1
 .....